+++ src/elpi_API.ml
@@ -35,8 +35,7 @@ let init ~builtins:(fname,decls) ~basedir:cwd argv =
    * program specific *)
   List.iter (function
     | Elpi_data.Builtin.MLCode (p,_) -> Elpi_data.Builtin.register p
-    | Elpi_data.Builtin.MLADT _ -> ()
-    | Elpi_data.Builtin.MLCData _ -> ()
+    | Elpi_data.Builtin.MLData _ -> ()
     | Elpi_data.Builtin.LPCode _ -> ()
     | Elpi_data.Builtin.LPDoc _ -> ()) decls;
   (* This is a bit ugly, since we print and then parse... *)
@@ -226,44 +225,24 @@ module Extend = struct
     exception No_clause = Elpi_data.No_clause
     include Elpi_data.Builtin
 
-    let adt { ADT.ty; constructors; doc } = 
-      let constructors = ADT.compile_constructors ty constructors in
-      {
-        ty;
-        doc;
-        readback = (ADT.readback ~look:Data.look ty constructors);
-        embed = (ADT.embed ty constructors);
-      }
+    let adt x = adt ~look:Data.look x
+    (* TODO: use typeabbrv instead of macro *)
+    let cdata ~name ?doc ?constants cd =
+      let prefix =
+        if name = "int" || name = "float" || name = "string" then ""
+        else "@" in
+      cdata ~look:Data.look ~name:(prefix^name) ?doc ?constants cd
 
-    let cdata ~name:ty ?(doc="") ?(constants=Data.Constants.Map.empty)
-      { CData.cin; isc; cout }
-    =
-      let ty = TyName ty in
-      let embed ~depth:_ _ _ state x =
-        state, Data.CData (cin x), [] in
-      let readback ~depth _ _ state t =
-        let module R = (val !r) in let open R in
-        match R.deref_head ~depth t with
-        | Data.CData c when isc c -> state, cout c
-        | Data.Const i as t when i < 0 ->
-            begin try state, Data.Constants.Map.find i constants
-            with Not_found -> raise (TypeErr(ty,t)) end
-        | t -> raise (TypeErr(ty,t)) in
-      { embed; readback; ty; doc }
 
     let int    = cdata ~name:"int" Elpi_data.C.int
     let float  = cdata ~name:"float" Elpi_data.C.float
     let string = cdata ~name:"string" Elpi_data.C.string
     let loc    = cdata ~name:"loc" Elpi_data.C.loc
 
-    (* TODO: use typeabbrv instead of macro *)
-    let cdata ~name ?doc ?constants cd =
-      cdata ~name:("@"^name) ?doc ?constants cd
-
     let poly ty =
       let embed ~depth:_ _ _ state x = state, x, [] in
       let readback ~depth _ _ state t = state, t in
-      { embed; readback; ty = TyName ty; doc = "" }
+      { embed; readback; ty = TyName ty; pp_doc = (fun fmt () -> ()) }
     let any = poly "any"
 
     let map_acc_embed f s l =
@@ -293,7 +272,7 @@ module Extend = struct
         let module R = (val !r) in let open R in
         map_acc_readback (d.readback ~depth h c) s (lp_list_to_list ~depth t)
       in
-      { embed; readback; ty = TyApp ("list",d.ty,[]); doc = "" }
+      { embed; readback; ty = TyApp ("list",d.ty,[]); pp_doc = (fun fmt () -> ()) }
 
 
     let builtin_of_declaration ~file_name x = file_name, x